Split IModifyRequets, fix xml comments
